% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datastepr.R
\docType{class}
\name{datastep}
\alias{datastep}
\title{An implementation of a data-step}
\format{An \code{\link{R6Class}} generator object}
\usage{
datastep
}
\description{
An implementation of a data-step
}
\section{Fields}{

  \describe{
    \item{\code{i}}{An index variable}}

  \describe{
    \item{\code{envir}}{An environment for evaluation}}

  \describe{
    \item{\code{results}}{A dataframe of results}}
}

\section{Methods}{

  \describe{\item{\code{set(df)}}{Import all values in row
    \code{self$i} of \code{df} into \code{self$env}}}

  \describe{\item{\code{lag_(.dots)}}{\code{\link{select_}(.dots)} from the last row of
    \code{self$results} and import into \code{self$env}}}

  \describe{\item{\code{lag(...)}}{\code{\link{select}(...)} from the last row of
    \code{self$results} and import into \code{self$env}}}

  \describe{\item{\code{output_(.dots)}}{Convert the environment to a dataframe,
    using only vectors, \code{\link{select_}(.dots)} and \code{\link{bind_cols}}
    onto \code{self$results}}}
}
\keyword{datasets}

