\name{datastepr-package}
\alias{datastepr-package}
\alias{datastepr}
\docType{package}
\title{
An implementation of a data step
}
\description{
Based on a SAS data step. Allows rowwise dynamic building of data.
}
\details{
\tabular{ll}{
Package: \tab datastepr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-06\cr
License: \tab CC0\cr
}
}
\author{
Brandon Taylor

Maintainer: Brandon Taylor <brandon.taylor221@gmail.com>
}