% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastepr.R
\name{toDf}
\alias{toDf}
\title{Append an object to a dataframe.}
\usage{
toDf(object, dataframe)
}
\arguments{
\item{object}{An object which can be coerced to a list (e.g. an environment)}

\item{dataframe}{A data frame}
}
\value{
An appended dataframe
}
\description{
Convert an object to a list, select only vector entries, coerce to a data.frame, 
and append to the given data frame.
}
\examples{
toDf(list(a = 1, b = 2, data.frame()), data.frame())
toDf(environment(), data.frame())
}

