% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastepr.R
\name{toEnv}
\alias{toEnv}
\title{Append an object to an environment}
\usage{
toEnv(object, environment)
}
\arguments{
\item{object}{An object which can be coerced to a list (e.g. an environment)}

\item{environment}{An environment}
}
\value{
An appended environment
}
\description{
A function to coerce an object to a list and append the list to an environment
}
\examples{
toEnv(data.frame(a = 1, b = 2), environment())
toEnv(list(a = 1, b = 2), environment())
toEnv(environment(), new.env())
}

